Allow ble_conn_params to still use APP_TIMER if needed. Particularly in the case of the Bootloader.
This is made optional based on a preprocessor macro: USE_APP_TIMER.
Some Nordic code assumes the use of AppTimer, and if we replace it with mbed's timer, then a lot more code gets pulled in than necessary. Ideally this should not have been a problem, but it is, and the bootloader is a bit unforgiving in its size constraints.