Add FOTA bootloader image
This file was previously stored in the mbed SDK, but it was decided that
putting it inside the nRF library would make more sense.

The exact location isn't too important, since mbed SDK is looking for
file names in all scanned resource files, when building a BOOT image.

It was built from the current ARMmbed dfu-bootloader repository.
It uses the internal LF clock source only, so it is compatible with all
nRF51822 devices. In addition, it doesn't rely on a specific SoftDevice:
the SoftDevice information struct is inspected at a fixed address, to
find out the application entry point.

As a temporary solution, we'll duplicate the file because mbed assumes
we need per-SoftDevice bootloaders. In the future, this will be a single
file.

In order to ease development, this bootloader pretends to use DFU version
0.4, so recent DFU applications won't mandate an init packet.