Add support for directed advertising mode
Implement the directed advertising mode (just pushed to the ble repo) for nRF51. I am not sure what the original
comment saying "This requires a proper security implementation, etc." means - as far as I can tell these are the only changes needed.

I have tested this by flashing it to an nRF51-DK, and scanning for the advertising packets with the Sniffer program. It found them ok.
nRF Connect doesn't seem to find them but I assume that is an Android limitation (or maybe scans aren't meant to report directed adverts).

This commit also includes a very minor bug fix - the minimum connection interval for ADV_SCANNABLE_UNDIRECTED is 100 ms.